Move main code into src folder
