+++ src/slash.js
